feat: 优化代码格式和逻辑，减少不必要的类型转换和增强可读性

+++ src/net/fpsboost/module/impl/SmokeCrosshair.java
@@ -154,7 +154,7 @@ public class SmokeCrosshair extends Module {
             this.curDynamicValues -= this.aerodynamicSpeedValue();
         }
         if (this.curDynamicValues < this.DynamicValues) {
-            this.curDynamicValues += (int) this.aerodynamicSpeedValue();
+            this.curDynamicValues += this.aerodynamicSpeedValue();
         }
         if (this.DynamicValues - this.curDynamicValues < this.aerodynamicSpeedValue() && this.DynamicValues - this.curDynamicValues > -this.aerodynamicSpeedValue()) {
             this.curDynamicValues = this.DynamicValues;
